Add GPU/CUDA implementation of the Mandelbrot set
This implementation exists in parallel to the CPU
implementation. The creation of the Mandelbrot set
is ~10 times faster than the CPU implementation.

No tests are added as there are some problems getting
them to run in debug and under valgrind. This is of
course not optimal and should be addressed asap.